objhits and Baddie: the hit pair, the sweep entry and the enemy state are pointers
objhits.c carried seventeen int-as-pointer parameter slots across eight
functions; every use in the bodies was already an explicit (GameObject*)
or (ObjAnimComponent*) cast, and anim sits at offset 0 of GameObject, so
the cast spellings become obj->anim.f and &obj->anim directly.  Closing
the parameters exposed the rest of the component: ObjHits_Update's five
object locals, its object list cursor, and ObjHitsSweepEntry.obj, which
stored a GameObject in an int field within a single translation unit.
ObjHitReact_Update's callers in six DLLs drop the (int) they had to add.

Baddie.c's `state` reaches EnemyState* in six functions and three more
locals.  Two of those bodies did byte arithmetic on the old integer:
Tricky_applyFloorResponse passed (void*)(state + 4) to the path-control
interface, which the header already documents as the record at +4, so it
is &state->flags; enemyObjAnimUpdate memcpy'd state + 0x2c4 from
state + 0x2b8, which are unk2C4 and lookDirX.  enemy_update and
enemy_SeqFn keep their u8*/char* state because they index it by byte
offset, and cast at the two call sites instead.

baddieSetMove is left alone: ten files under objects/202 call it raw.

Byte-neutral: 1005 objects compared against a pristine build at the same
parent, 0 differ, 2 renumbered (literal-pool @NNN only).  99.817320 /
2517908 / 1195969 / 914 of 1043 unchanged.

+++ src/dlls/objects/417/417.c
@@ -684,7 +684,7 @@ void NW_mammoth_update(GameObject* obj, int unusedArg) {
             hitReactEntries = &tables[0]->normalHitReactEntry;
         }
         state->hitReactState =
-            ObjHitReact_Update((int)obj, hitReactEntries, 1, state->hitReactState, &state->hitReactStepScale);
+            ObjHitReact_Update(obj, hitReactEntries, 1, state->hitReactState, &state->hitReactStepScale);
         if (state->hitReactState != 0) {
             characterHeadLookRelax(obj, &state->eyeAnim);
             characterDoEyeAnims(obj, &state->eyeAnim);
